requirements.txt
PyYAML==5.4.1
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==1.5
atomicwrites==1.4.0
attrs==21.1.0
by==0.0.2
colorama==0.4.4
execnet==1.8.0
importlib-metadata==4.0.1
iniconfig==1.1.1
packaging==20.9
pip==21.1.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-xdist==2.2.1
selenium==3.141.0
setuptools==56.1.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
